Encapsulate almost all Ref<Object> inside Value.
Right now, Finch burns a lot of cycles on just ref-count thrashing. This is
mainly because we pass around Ref<Object>s by *value* even when it's not
needed (like for args that don't store the object).

It also has a really naive object representation. This is the first step
towards fixing this and also gets rid of a bunch of ref-count thrashes. It
introduces a Value class that is a Finch value (or a reference to one, the
distinction is a little fuzzy right now). We pass those by reference when
possible to avoid copying.

Eventually, Value will also be a more tightly packed object representation
instead of just wrapping a Ref<Object> like it does now. This will let us
do stuff like unboxed numbers.